add auth details - VI
